Eliminate remaining circularity!
